# opt: skip space use shuffle SIMD instruction