performance:

```
cpu: Intel(R) Xeon(R) Platinum 8260 CPU @ 2.40GHz
                                      │ /tmp/tmphc12qb39.source.txt │    /tmp/tmp6sv5jjce.target.txt     │
                                      │           sec/op            │   sec/op     vs base               │
Decoder_Generic_Sonic-8                                 278.0µ ± 4%   264.3µ ± 3%  -4.93% (p=0.001 n=10)
Decoder_Generic_Sonic_Fast-8                            211.6µ ± 4%   205.1µ ± 4%  -3.09% (p=0.023 n=10)
Decoder_Binding_Sonic-8                                105.03µ ± 4%   99.42µ ± 3%  -5.35% (p=0.000 n=10)
Decoder_Binding_Sonic_Fast-8                            83.17µ ± 5%   75.23µ ± 4%  -9.54% (p=0.000 n=10)
Decoder_Parallel_Generic_Sonic-8                        56.05µ ± 3%   54.35µ ± 2%  -3.04% (p=0.001 n=10)
Decoder_Parallel_Generic_Sonic_Fast-8                   43.87µ ± 4%   43.01µ ± 4%       ~ (p=0.247 n=10)
Decoder_Parallel_Binding_Sonic-8                        20.84µ ± 8%   19.50µ ± 7%  -6.44% (p=0.005 n=10)
Decoder_Parallel_Binding_Sonic_Fast-8                   17.20µ ± 5%   15.61µ ± 3%  -9.27% (p=0.000 n=10)
geomean                                                 67.92µ        64.20µ       -5.49%
```